Add files via upload
census_id and GEOID_10 are the same, we use ArcMap (part of the ArcGIS products) to join the count to the census tract shapefiles to display the count data. Please feel free to use whatever mapping api you'd like and let me know if there is anything we can do to make mapping easier for you all.